requirements.txt
PyGithub
pytest
pytest-timeout
tabulate